Fix up naming of not_null local in Parser impl for ColumnSpec
The definiton of ColumnSpec has a field 'not_null', but the parser
was using a local called 'nullable'. It still did the right thing, it
was just confusing.

Testing: The tests are busted :( but I checked before and after this
patch that non-key columns with spec NOT NULL and without are correctly
created as NOT NULL and NULLABLE, respectively, by checking the Kudu
master web UI.